[+BUGFIX] Activated Orphan Removal for Person -> PersonName
A Person has a PersonName; but it is *very* unlikely that one uses the same
PersonName in two Person instances.

Generally, I feel that later it makes more sense if PersonName gets
a Value Object again, but we need this change for the conference site.

See http://www.doctrine-project.org/docs/orm/2.0/en/reference/working-with-associations.html#orphan-removal for an explanation of this feature.

Change-Id: Idbd7bd4b193a280a6c69ae4a058143a39e5ba549