[BUGFIX] Adjust styling and reposition form buttons
With this change the setup uses the default markup for modals
see http://twitter.github.com/bootstrap/javascript.html#modals

That comes with more space for the content and some clever
floating / responsive design tricks for smaller screens.

Besides this slightly adjusts the order of buttons to be more
intutive ("back" on the bottom left, "skip" and "next" on the
botton right).

Lastly this changes the setup form preset name from "setup" to
"typo3.setup" to avoid conflicts with 3rd party code.

Change-Id: I2b0e142daba193758941e528ef88964543303e07
Fixes: #38856
Fixes: #38857