[BUGFIX] A database change needs a rebuild of the policy roles cache
If the database name is changed (or a database is dropped and
re-created) the setup leaves an inconsistent state with the policy roles
cache not being in sync with the database.

It's safe to assume that after the database step the policy service
caches can be reset to force a rebuild in the next request.

Change-Id: I1034f06fc02e6d449e30c59c7739a591d459bc1e