[BUGFIX] Also check PHP_BINARY before using it
To make sure the binary defined in PHP_BINARY is actually a cli
binary we should check it as well. Otherwise it could also be a
CGI binary that will result in problems.

To avoid infinite loops with CGI we need to remove the
``SCRIPT_FILENAME`` environment variable as otherwise PHP will
execute that (which results again in a call of this code and again
and again and again and again and again....).
The value is set again afterwards to make sure no side effects occur.

Change-Id: I7963bbfcf6be7f766d1e89e9d926fdb850562284
Releases: master, 1.1
Resolves: NEOS-1011
Reviewed-on: http://review.typo3.org/37664
Reviewed-by: Karsten Dambekalns <karsten@typo3.org>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Aske Ertmann <aertmann@gmail.com>
Tested-by: Aske Ertmann <aertmann@gmail.com>