[BUGFIX] Catch exceptions thrown during form rendering (finishers)
In addition to catching exception for the step postprocessors,
exceptions are caught when rendering the form. This allows form
finishers to display nicely formatted error messages without having
to be presented on the next form rendering as it is only possible
currently.

Additionally changes the severity of flash messages generated from
setup exceptions from ``SEVERITY_WARNING`` to ``SEVERITY_ERROR``.

Change-Id: Ifbba8d4c45b87a6ec171f3fe72bbc093c7b93722
Releases: master, 1.1
Related: NEOS-758
Reviewed-on: http://review.typo3.org/38709
Reviewed-by: Aske Ertmann <aertmann@gmail.com>
Tested-by: Aske Ertmann <aertmann@gmail.com>
Reviewed-by: Bastian Waidelich <bastian@typo3.org>