BUGFIX: Adjust index names to match Doctrine DBAL 2.5
The use of Doctrine 2.5 (instead of 2.4) exposes the fact that some
(old) index names in the Party database schema do not match the names
that are generated currently.

This adjusts those index names, something that is a one-time adjustment.

FLOW-427 #close Adjusts indexes as needed
FLOW-222 #close Adjusts indexes as needed