BUGFIX: Prevet stale object references in party cache
Holding a reference to an doctrine object can be dangerous
if something happens to the UOW during that time, so that
a new instance of the entity is produced, resulting in a
stale entity reference that must be considered detached.
We therefore should only cache the persistence identifier of
a connected party for easier retrieval. If said party was
already loaded by doctrine it will be available in the
identity map and should not trigger another query, making this
not less efficient than the original solution.